Callbacks from native to managed code must call static functions on AOT platforms.
In order to map these static function calls to an instance it is necessary to store a reference to the managed wrapper from the native wrapper and pass the reference as a parameter to the static function call. This changes implement this.

Also added four more functions in MultiBodyTree_wrap.